Module horrorshow::prelude

Traits that should always be imported.

Structs

FnRenderer
A template renderer. The html! {} macro returns a FnRenderer.

TemplateBuffer
A template buffer. This is the type that gets passed to closures inside templates.

Traits

Render
Something that can be rendered by reference.

RenderBox
Something that can be rendered once out of a box.

RenderMut
Something that can be rendered by mutable reference.

RenderOnce
Something that can be rendered once.

Template
A template that can be rendered into something.