Trait horrorshow::Render

pub trait Render: RenderMut {
    fn render_tmpl<'a>(&self, tmpl: &mut TemplateBuilder<'a>);

    fn render(&self) -> String { ... }
    fn render_string(&self, string: &mut String) { ... }
    fn render_fmt(&self, writer: &mut Write) -> Result<(), Error> { ... }
    fn render_io(&self, writer: &mut Write) -> Result<(), Error> { ... }
}

A component that can be appended to a template.

Required Methods

fn render_tmpl<'a>(&self, tmpl: &mut TemplateBuilder<'a>)

Render this into a template builder.

Provided Methods

fn render(&self) -> String

Render this into a new String.

fn render_string(&self, string: &mut String)

Render this into an existing String.

Note: You could also use render_into_fmt but this is noticeably faster.

fn render_fmt(&self, writer: &mut Write) -> Result<(), Error>

Render this into something that implements fmt::Write.

Renderer also implements Display but that's about twice as slow...

fn render_io(&self, writer: &mut Write) -> Result<(), Error>

Render this into something that implements io::Write.

Note: If you're writing directly to a file/socket etc., you should seriously consider wrapping your writer in a BufWriter. Otherwise, you'll end up making quite a few unnecessary system calls.