Struct horrorshow::prelude::TemplateBuffer

pub struct TemplateBuffer<'a> {
    // some fields omitted
}

A template buffer. This is the type that gets passed to closures inside templates.

Example:

    html! {
        |tmpl /*: &mut TemplateBuffer */| tmpl << "Some String";
    };

Methods

impl<'a> TemplateBuffer<'a>

fn record_error<E: Into<Box<Error + Send + Sync>>>(&mut self, e: E)

fn write_raw(&mut self, text: &str)

Write a raw string to the template output.

fn write_fmt(&mut self, args: Arguments)

Escape and write the formatted arguments to the template output.

Example:

write!(tmpl, "{} + {}", 0, 1);

fn write_str(&mut self, text: &str)

Escape and write a string to the template output.

fn as_writer<'b>(&'b mut self) -> TemplateWriter<'a, 'b>

Returns an escaping Write adapter.

fn as_raw_writer<'b>(&'b mut self) -> RawTemplateWriter<'a, 'b>

Returns a non-escaping Write adapter.